Added options, consolidated and removed code, use of OSMDroid and OSMBonus
When you select a station, a button is enabled to allow the user to use that as their destination. Also, there is a small map that gives the user can use to barely compare their location and the station highlighted.
Using OSMDroid bonus, we give you directions to the station and not the entrance in a map format. We have yet to change it to entrances, but will in next commit.
Eliminate API calls by only selecting stations that are in the same state as the user (if they’re in NY or NJ). This will give us some time to call all the entrances of the three closest stations in order to give the user an accurate measure of distance or duration.
Distance/Duration from the origin is now judged not by the location of the station, but by the closest entrance of each of the three closest stations
The possibility now exists that there will be less than three stations appearing in the results. This is expected and is handled.
Some text corrections
Removed confirmOptions in ClosestStatiosnActivity. No longer will the options the user selected be displayed.
Partly removed the need for ObjectSort, especially in ClosestStationsIntent. However it is still needed in ClosestStatoinsActivity
The closest entrance at each of the three stations is obtainable, but nothing is done with it yet. It is only passed to the second activity.
Added to README credit to OSMDroid and OSMBonusPack
Removed “shea” folder from src
Added the steps needed to convert a manual location to their geo-coordinates, as well as the steps needed to download and parse information to find the three closest stations, onto their own separate threads. This will eliminate the issue of skipping frames between the two first activities.
An alert is given to the user when they enter a manual location and it cannot be converted to geo-coordinates
The location of the user is now obtained in the first activity where options are selected. Also, we use the last known location of GPS or Network if it is obtainable. If not, then we enable a GPS and Network listener. They're disabled when they have a location, and again when the first intent is being created. Last known locations are good if they're within 50 meters and were obtained in the last two minutes.
Added handicap options for stations and entrances. Now if the user selects elevator or escalator, we only search for stations and entrances that are accessible in one of those ways.