Near-final version of app
This will serve as my final commit before I release my application to the Google Play Store. A lot code was added/removed/changed between now and the last committ. I'll therefore be brief. The main activity was split into two fragments, one for the closest station search and another for a system map, which includes the weekday and nights/weekend map. The main activity also includes a tab for both fragments, as well as a link to a donate page. Ads have also been placed under the tabs in the action bar, as well as at the end of the page where you select which station you want directions to. All of this also takes into account the closing of the World Trade Center and Exchange Place station on weekends throughout 2014. The tabs will allow for expansion of features for the PATH system, freeing the navigation bar (that I'll add in time) available for selected what transit system you'll like to use.